datacol/requirements.txt
absl-py==0.8.1
alembic==1.3.1
asgiref==3.2.3
astor==0.8.0
astroid==2.3.3
autopep8==1.4.4
bcrypt==3.1.4
beautifulsoup4==4.8.1
blinker==1.4
bootstrap4==0.1.0
bs4==0.0.1
certifi==2019.3.9
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.4.3
confusable-homoglyphs==3.0.0
cycler==0.10.0
decouple==0.0.7
defusedxml==0.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.4
django-allauth==0.39.1
django-autofixture==0.12.1
django-bootstrap==0.2.4
django-bootstrap3==9.1.0
django-bootstrap4==1.1.1
django-braces==1.14.0
django-cleanup==4.0.0
django-cors-headers==3.2.1
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-elasticsearch-dsl==7.1.1
django-forms-bootstrap==3.1.0
django-registration==2.4.1
django-tinymce==2.7.0
djangorestframework==3.11.0
docx2txt==0.8
easy-thumbnails==2.7
elasticsearch==7.0.5
elasticsearch-dsl==7.1.0
Flask==1.0
Flask-Admin==1.5.4
Flask-Bcrypt==0.7.1
Flask-Debug==0.4.3
Flask-Ext==0.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-MongoAlchemy==0.7.2
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.3.2
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.2
fulltext==0.5
gast==0.3.2
get==2019.4.13
grpcio==1.25.0
gunicorn==19.7.1
h5py==2.10.0
heroku==0.1.4
idna==2.8
inflection==0.3.1
isort==4.3.21
itsdangerous==0.24
Jinja2==2.10
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
lxml==4.4.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.0
matplotlib==3.1.1
mccabe==0.6.1
MongoAlchemy==0.19
nltk==3.4.5
numpy==1.17.4
oauthlib==3.0.1
pandas==0.25.3
pep8==1.7.1
Pillow==7.0.0
post==2019.4.13
protobuf==3.10.0
psycopg2==2.8.4
public==2019.4.13
pycodestyle==2.5.0
pycparser==2.18
pylint==2.4.4
pymongo==2.8.1
pyparsing==2.4.5
python-dateutil==2.8.1
python-decouple==3.1
python-editor==1.0.4
python-engineio==3.10.0
python-pptx==0.6.18
python-socketio==4.3.1
python3-openid==3.1.0
pytz==2018.3
PyYAML==5.1.2
query-string==2019.4.13
request==2019.4.13
requests==2.21.0
requests-oauthlib==1.2.0
six==1.14.0
soupsieve==1.9.4
SQLAlchemy==1.2.7
sqlparse==0.2.4
static3==0.7.0
stripe==2.27.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
textblob==0.15.3
tornado==6.0.3
typed-ast==1.4.0
unicode==2.7
Unipath==1.1
urllib3==1.24.2
virtualenv==16.7.9
Werkzeug==0.14.1
whitenoise==3.3.1
Whoosh==2.7.4
wrapt==1.11.2
WTForms==2.0
xlrd==1.2.0
XlsxWriter==1.2.2
